Restructured folders
added features and layout folder
