--- src/components/PastEvents.tsx
@@ -1,3 +1,0 @@
-function PastEvents() {
-  return <div></div>;
-}
